requirements.txt
django
pillow
gunicorn
whitenoise
cloudinary